Fix tests on Windows
